+++ mycodo/mycodo_flask/utils/utils_dashboard.py
@@ -78,6 +78,7 @@ def dashboard_add(form_base, form_object, display_order):
         new_graph.enable_navbar = form_object.enable_navbar.data
         new_graph.enable_rangeselect = form_object.enable_range.data
         new_graph.enable_export = form_object.enable_export.data
+        new_graph.enable_graph_shift = form_object.enable_graph_shift.data
         new_graph.enable_manual_y_axis = form_object.enable_manual_y_axis.data
         new_graph.y_axis_min = form_object.y_axis_min.data
         new_graph.y_axis_max = form_object.y_axis_max.data
@@ -238,6 +239,7 @@ def dashboard_mod(form_base, form_object, request_form):
         mod_graph.enable_navbar = form_object.enable_navbar.data
         mod_graph.enable_export = form_object.enable_export.data
         mod_graph.enable_rangeselect = form_object.enable_range.data
+        mod_graph.enable_graph_shift = form_object.enable_graph_shift.data
         mod_graph.enable_manual_y_axis = form_object.enable_manual_y_axis.data
         mod_graph.y_axis_min = form_object.y_axis_min.data
         mod_graph.y_axis_max = form_object.y_axis_max.data